Changed typedef for EGLTimeKHR to use already-defined uint64_t instead of unsigned long long which is not supported in some compilers (e.g. MSVC 6)
git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@1963 232f8b59-042b-4e1e-8c03-345bb8c30851

+++ make/stub_includes/egl/EGL/eglext.h
@@ -278,7 +278,7 @@ EGLAPI EGLBoolean EGLAPIENTRY eglNvmaOutputSurfacePbufferUnlock(EGLDisplay displ
 #define EGL_KHR_sync 1
 typedef void* EGLSyncKHR;
 typedef void* NativeSyncKHR;
-typedef unsigned long long EGLTimeKHR;
+typedef uint64_t EGLTimeKHR;
 #ifdef EGL_EGLEXT_PROTOTYPES
 EGLSyncKHR eglCreateFenceSyncKHR( EGLDisplay dpy, EGLenum condition, const EGLint *attrib_list ); 
 NativeSyncKHR eglCreateNativeSyncKHR( EGLSyncKHR sync );
